extracting surveys and questions step
