+++ src/app/components/admin-new-trial/survey-questions.step.component.html
@@ -1,5 +1,5 @@
 <div class='mat-headline'>Survey Settings and Questions</div>
-<mat-accordion [formGroup]='parentForm'>
+<mat-accordion [formGroup]='parentForm' *ngIf='surveys'>
 	<mat-expansion-panel
 		formArrayName='surveys'
 		*ngFor='let survey of surveys.controls; index as s'
@@ -7,14 +7,14 @@
 		<mat-expansion-panel-header>
 			<mat-panel-title>
 				<mat-icon *ngIf="!survey.get('survey_name').value">add_circle</mat-icon>
-				<div class='survey'>{{ survey.get('survey_name').value || 'add new survey' }}</div>
+				<div class='survey-title'>{{ survey.get('survey_name').value || 'add new survey' }}</div>
 			</mat-panel-title>
 			<mat-panel-description>
 				{{ (survey.get('survey_questions').value.length > 1) ? (survey.get("survey_questions").value.length-1) + " question"+(survey.get("survey_questions").value.length>2?"s":"") : "" }}
 			</mat-panel-description>
 		</mat-expansion-panel-header>
 		<ng-template matExpansionPanelContent>
-			<div [formGroupName]='s' class='col'>
+			<div [formGroupName]='s' class='col section1'>
 				<mat-form-field>
 					<input
 						matInput
@@ -22,19 +22,9 @@
 						placeholder='Survey Name:'
 						name='survey_name'
 						formControlName='survey_name'
-						(input)='changeInput($event)'/>
+						(input)='changeSurveyName($event)'/>
 				</mat-form-field>
 				<div *ngIf='survey.get("survey_name").value.length > 0'>
-					<div class='mat-subheading-2'>Request survey from the following Groups:</div>
-					<div class='item_group_cell'
-						*ngFor='let group of groups.controls; index as t'>
-						<mat-checkbox
-							type='checkbox'
-							name='survey_groups'
-							(input)='changeGroupAssignment($event, survey.controls["survey_groups"], group.controls["group_id"].value )'>
-							{{ group.controls['group_name'].value || 'Group '+(group.controls['group_id'].value+1) }}
-						</mat-checkbox>
-					</div>
 					<div class='mat-subheading-2'>Request survey at these intervals:</div>
 					<div class='row'>
 						<div class='col'>
@@ -66,6 +56,16 @@
 							</mat-form-field>
 						</div>
 					</div>
+					<div class='mat-subheading-2'>Request survey from the following Groups:</div>
+					<div class='col'>
+						<mat-checkbox
+							*ngFor='let group of groups.controls; index as t'
+							type='checkbox'
+							name='survey_groups'
+							(input)='changeGroupAssignment($event, survey.controls["survey_groups"], group.controls["group_id"].value )'>
+							{{ group.controls['group_name'].value || 'Group '+(group.controls['group_id'].value+1) }}
+						</mat-checkbox>
+					</div>
 				</div>
 			</div>
 			<mat-divider *ngIf='survey.get("survey_name").value.length > 0'></mat-divider>
@@ -74,19 +74,19 @@
 				<div class='mat-h2'>Questions:</div>
 				<div formArrayName='survey_questions'
 					*ngFor='let question of survey.controls["survey_questions"].controls; index as q'
-					class='col'>
+					class='col question'>
 					<div [formGroupName]='q' class='row'>
-						<h3 class='mat-h3 formFieldLabel'>№{{q+1}}:</h3>
-						<mat-form-field class='fullFlex'>
+						<h3 class='mat-h3 form-field-label'>№{{q+1}}:</h3>
+						<mat-form-field class='full-flex'>
 							<input
 								matInput
 								type='text'
 								placeholder='Question Text:'
 								name='question_text'
 								formControlName='question_text'
-								(input)='changeInput($event)'/>
+								(input)='changeQuestionText($event)'/>
 						</mat-form-field>
-						<mat-form-field class='fullFlex'>
+						<mat-form-field class='full-flex'>
 							<mat-select
 								name='question_type'
 								placeholder='Question Type:'
@@ -97,8 +97,8 @@
 						</mat-form-field>
 					</div>
 					<div [formGroupName]='q' class='row'
-					*ngIf='question.get("question_type").value !== "text"'>
-						<mat-form-field class='fullFlex h3Indent'>
+						*ngIf='question.get("question_type").value !== "text"'>
+						<mat-form-field class='full-flex indent'>
 							<input
 								matInput
 								type='text'
@@ -106,7 +106,7 @@
 								name='question_options'
 								formControlName='question_options'/>
 							<mat-hint align='start'>Delimit the options with the pipe symbol (|)</mat-hint>
-								<mat-hint align='end'></mat-hint>
+							<mat-hint align='end'></mat-hint>
 						</mat-form-field>
 					</div>
 					<mat-divider *ngIf='survey.controls["survey_questions"].controls.length-1 !== q'></mat-divider>
